chore: split admin command up into multiple files.
